🗃️ ugly_percent 타입명 변경

+++ src/main/java/com/bbogle/yanu/domain/farm/domain/FarmEntity.java
@@ -52,7 +52,7 @@ public class FarmEntity {
     @Min(0)
     @Max(100)
     @Column
-    private byte ugly_percent;
+    private float ugly_percent;
 
     @JsonIgnore
     @OneToMany(mappedBy = "farm", cascade = {CascadeType.REFRESH, CascadeType.DETACH, CascadeType.MERGE, CascadeType.PERSIST}, orphanRemoval = true)
