+++ src/main/java/com/bbogle/yanu/domain/farm/dto/RegisterFarmRequestDto.java
@@ -31,7 +31,7 @@ public class RegisterFarmRequestDto {
                 .email(email)
                 .address(address)
                 .geography(geography)
-                .ugly_percent((byte) 13)
+                .ugly_percent(13)
                 .build();
     }
 }
